+++ src/controllers/MainController.php
@@ -25,7 +25,9 @@ class MainController extends Controller
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
-                    //'logout' => ['post'],
+                    //'index' => ['get'],
+                    //'atom' => ['get'],
+                    //'rss' => ['get'],
                 ],
             ],
         ];
@@ -51,12 +53,12 @@ class MainController extends Controller
     {
         $response = Yii::$app->response;
 
-        $pages = XmlGeneratorModule::getInstance()->pagesArray;
+        $pages = XmlGeneratorModule::getInstance()->sitemapInstance;
 
         $response->format = Response::FORMAT_RAW;
         $response->headers->add('Content-Type', 'text/xml');
         $response->data = $this->renderPartial('sitemapXml', [
-            'items' => $pages
+            'sitemap' => $pages
         ]);
     }
 
@@ -68,13 +70,12 @@ class MainController extends Controller
     {
         $response = Yii::$app->response;
 
-        $atomItems = XmlGeneratorModule::getInstance()->atomArray;
+        $atomFeed = XmlGeneratorModule::getInstance()->atomInstance;
 
         $response->format = Response::FORMAT_RAW;
         $response->headers->add('Content-Type', 'text/xml');
         $response->data = $this->renderPartial('atomXml', [
-            'main' => $atomItems['main'] ?? [],
-            'items' => $atomItems['items'] ?? []
+            'feed' => $atomFeed
         ]);
     }
 
@@ -86,12 +87,12 @@ class MainController extends Controller
     {
         $response = Yii::$app->response;
 
-        $rssItems = XmlGeneratorModule::getInstance()->rssArray;
+        $rssChannel = XmlGeneratorModule::getInstance()->rssInstance;
 
         $response->format = Response::FORMAT_RAW;
         $response->headers->add('Content-Type', 'text/xml');
         $response->data = $this->renderPartial('rssXml', [
-            'items' => $rssItems
+            'channel' => $rssChannel
         ]);
     }
 }