+++ src/views/main/atomXml.php
@@ -1,14 +1,15 @@
 <?php echo '<?xml version="1.0" encoding="UTF-8"?>' . PHP_EOL ?>
 <feed xmlns="http://www.w3.org/2005/Atom">
-    <title><?= $main->title ?></title>
-    <link href="<?= $main->link ?>"/>
-    <updated><?= $main->updated ?></updated>
+    <?php if ( $feed ) {?>
+    <title><?= $feed->title ?></title>
+    <link href="<?= $feed->link ?>"/>
+    <updated><?= $feed->updated ?></updated>
     <author>
-        <name><?= $main->author ?></name>
+        <name><?= $feed->author ?></name>
     </author>
-    <id><?= $main->id ?></id>
+    <id><?= $feed->id ?></id>
 
-    <?php foreach ($items as $item): ?>
+    <?php foreach ($feed->items as $item): ?>
     <entry>
         <title><?= $item->title?></title>
         <link href="<?= $item->link?>"/>
@@ -17,4 +18,5 @@
         <summary><?= $item->summary?></summary>
     </entry>
     <?php endforeach; ?>
+    <?php } ?>
 </feed>