+++ src/views/main/sitemapXml.php
@@ -1,7 +1,8 @@
 <?php echo '<?xml version="1.0" encoding="UTF-8"?>' . PHP_EOL ?>
 <?php /*echo '<?xml-stylesheet type="text/css" href="/css/doc.css"?>' . PHP_EOL */?>
+<?php if ( $sitemap ) { ?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-<?php foreach ($items as $item): ?>
+<?php foreach ($sitemap->items as $item): ?>
     <url>
         <loc><?= $item->loc ?></loc>
         <lastmod><?= $item->lastmod ?></lastmod>
@@ -10,3 +11,4 @@
     </url>
 <?php endforeach; ?>
 </urlset>
+<?php } ?>